Refactor DNA match parser into separate module
