+++ gramps_webapi/api/resources/dna.py
@@ -23,7 +23,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Union
+from typing import Any
 
 from flask import abort
 from gramps.gen.const import GRAMPS_LOCALE as glocale
@@ -34,8 +34,9 @@ from gramps.gen.relationship import get_relationship_calculator
 from gramps.gen.utils.grampslocale import GrampsLocale
 from webargs import fields, validate
 
+from gramps_webapi.api.dna import parse_raw_dna_match_string
 from gramps_webapi.api.people_families_cache import CachePeopleFamiliesProxy
-from gramps_webapi.types import Handle, ResponseReturnValue
+from gramps_webapi.types import Handle, MatchSegment, ResponseReturnValue
 
 from ...types import Handle
 from ..util import get_db_handle, get_locale_for_language, use_args
@@ -45,8 +46,6 @@ from .util import get_person_profile_for_handle
 SIDE_UNKNOWN = "U"
 SIDE_MATERNAL = "M"
 SIDE_PATERNAL = "P"
-
-Segment = dict[str, Union[float, int, str]]
 
 
 class PersonDnaMatchesResource(ProtectedResource):
@@ -195,7 +194,7 @@ def get_match_data(
 
 def get_segments_from_note(
     db_handle: DbReadBase, handle: Handle, side: str | None = None
-) -> list[Segment]:
+) -> list[MatchSegment]:
     """Get the segements from a note handle."""
     try:
         note: Note | None = db_handle.get_note_from_handle(handle)
@@ -204,73 +203,23 @@ def get_segments_from_note(
     if note is None:
         return []
     raw_string: str = note.get()
-    return parse_raw_dna_match_string(raw_string, side=side)
+    return parse_raw_match_string_with_default_side(raw_string, side=side)
 
 
-def parse_raw_dna_match_string(
+def parse_raw_match_string_with_default_side(
     raw_string: str, side: str | None = None
-) -> list[Segment]:
-    """Parse a raw DNA match string and return a list of segments."""
+) -> list[MatchSegment]:
+    """Parse a raw DNA match string and return a list of segments.
+
+    If the side is unknown, optionally set it to a default value.
+    """
+    original_segments = parse_raw_dna_match_string(raw_string)
+    if side is None:
+        return original_segments
     segments = []
-    for line in raw_string.split("\n"):
-        data = parse_line(line, side=side)
-        if data:
-            segments.append(data)
-    return segments
-
-
-def parse_line(line: str, side: str | None = None) -> Segment | None:
-    """Parse a line from the CSV/TSV data and return a dictionary."""
-    if "\t" in line:
-        # Tabs are the field separators. Now determine THOUSEP and RADIXCHAR.
-        # Use Field 2 (Stop Pos) to see if there are THOUSEP there. Use Field 3
-        # (SNPs) to see if there is a radixchar
-        field = line.split("\t")
-        if "," in field[2]:
-            line = line.replace(",", "")
-        elif "." in field[2]:
-            line = line.replace(".", "")
-        if "," in field[3]:
-            line = line.replace(",", ".")
-        line = line.replace("\t", ",")
-    field = line.split(",")
-    if len(field) < 4:
-        return None
-    chromo = field[0].strip()
-    start = get_base(field[1])
-    stop = get_base(field[2])
-    try:
-        cms = float(field[3])
-    except (ValueError, TypeError, IndexError):
-        return None
-    try:
-        snp = int(field[4])
-    except (ValueError, TypeError, IndexError):
-        snp = 0
-    seg_comment = ""
-    side = side or SIDE_UNKNOWN
-    if len(field) > 5:
-        if field[5] in {SIDE_MATERNAL, SIDE_PATERNAL, SIDE_UNKNOWN}:
-            side = field[5].strip()
+    for segment in original_segments:
+        if segment["side"] == SIDE_UNKNOWN:
+            segments.append(segment | {"side": side})
         else:
-            seg_comment = field[5].strip()
-    return {
-        "chromosome": chromo,
-        "start": start,
-        "stop": stop,
-        "side": side,
-        "cM": cms,
-        "SNPs": snp,
-        "comment": seg_comment,
-    }
-
-
-def get_base(num: str) -> int:
-    """Get the number as int."""
-    try:
-        return int(num)
-    except (ValueError, TypeError):
-        try:
-            return int(float(num) * 1000000)
-        except (ValueError, TypeError):
-            return 0
+            segments.append(segment)
+    return segments
